% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{adjust_baseline}
\alias{adjust_baseline}
\title{adjust model run in absence of vaccination so diagnoses are
spread across vaccine protected and non-vaccine protected strata as if
the rate of movement was the same as in an equivalent model run in the
presence of vaccination}
\usage{
adjust_baseline(baseline, y)
}
\arguments{
\item{baseline}{A model run in the absence of vaccine uptake}

\item{y}{A model run in the presence of vaccine uptake}
}
\value{
An adjusted baseline in which 'a' and 's' diagnoses in the baseline
are divided across vaccine protected and non-vaccine protected strata
allowing comparison of model runs with of baselines by vaccine-status
}
\description{
adjust model run in absence of vaccination so diagnoses are
spread across vaccine protected and non-vaccine protected strata as if
the rate of movement was the same as in an equivalent model run in the
presence of vaccination
}
