% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{aggregate}
\alias{aggregate}
\title{aggregate model runs by vaccination group x vaccine strata}
\usage{
aggregate(
  x,
  what,
  as_incid = FALSE,
  stratum = NULL,
  f = identity,
  stochastic = FALSE,
  ...
)
}
\arguments{
\item{x}{a transformed model run output}

\item{what}{a character string of the cumulative trajectory to transform}

\item{as_incid}{logical specifying whether to convert to incidence}

\item{stratum}{an integer denoting the stratum to aggregate over, if NULL
aggregation is over all strata}

\item{f}{function to apply across parameter sets (eg. mean / median)}

\item{stochastic}{indicates if function is being used on output of
stochastic or deterministic version of the trial model, only matters if
as_incid = TRUE}

\item{...}{named arguments to pass to f}
}
\value{
a transformed time series / array
}
\description{
aggregate model runs by vaccination group x vaccine strata
}
