% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{compare}
\alias{compare}
\title{Calculate the log likelihood of the data given the parameters
diagnoses and attendances lhoods are negative binomial
p_symp lhood is betabinomial}
\usage{
compare(pars, transform)
}
\arguments{
\item{pars}{A named vector of parameters}

\item{transform}{the transform function to use in the comparison}
}
\value{
a single log likelihood
}
\description{
Calculate the log likelihood of the data given the parameters
diagnoses and attendances lhoods are negative binomial
p_symp lhood is betabinomial
}
