% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_diagnosis_waning_map}
\alias{create_diagnosis_waning_map}
\title{Create mapping for movement between strata due to diagnosis waning}
\usage{
create_diagnosis_waning_map(n_vax, z, n_diag_rec = 1)
}
\arguments{
\item{n_vax}{Integer in (0, 5) denoting total number of strata}

\item{z}{Scalar denoting rate of waning diagnosis}

\item{n_diag_rec}{integer for the number of diagnosis history substrata}
}
\value{
an array of the mapping
}
\description{
Create mapping for movement between strata due to diagnosis waning
}
