% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_uptake_map}
\alias{create_uptake_map}
\title{Creates uptake mapping array with dimensions n_group x n_vax x n_vax
and assigns the relevant primary uptake and booster uptake values defined
by the user.}
\usage{
create_uptake_map(
  n_group,
  n_vax,
  primary_uptake,
  booster_uptake,
  i_eligible,
  i_v
)
}
\arguments{
\item{n_group}{scalar indicating number of activity groups}

\item{n_vax}{scalar indicating number the number of stratum in the model}

\item{primary_uptake}{proportion of the unvaccinated population who accept
primary vaccination}

\item{booster_uptake}{proportion of the formerly fully vaccinated, waned
population who accept a booster vaccination dose}

\item{i_eligible}{vector of indices of stratum which are eligible for 
vaccination, of same length as the number of paths from unvaccinated to
vaccinated}

\item{i_v}{vector of indices of stratum which are vaccinated and experience
protection}
}
\value{
an array of the uptakes with dimensions n_group x n_vax x n_vax
}
\description{
Creates uptake mapping array with dimensions n_group x n_vax x n_vax
and assigns the relevant primary uptake and booster uptake values defined
by the user.
}
