% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xpvwrh.R
\name{create_uptake_map_xpvwrh}
\alias{create_uptake_map_xpvwrh}
\title{Creates uptake mapping for the branching XPVWRH model where
individuals can move from unvaccinated (X) to vaccinated (V) or partially
vaccinated (P) as well as revaccinated from waned (W) to (R) and, and
partially vaccinated (P) to fully vaccianted (V). The former
reflects the specific indices which are chosen for assigning uptakes.}
\usage{
create_uptake_map_xpvwrh(array, r1, r2, r2_p, booster_uptake, n_erlang = 1)
}
\arguments{
\item{array}{a vaccine map array of dimensions n_group by n_vax by n_vax
generated through create_vax_map_branching()}

\item{r1}{proportion of population offered vaccine only accepting the first
dose, becoming partially vaccinated}

\item{r2}{proportion of the population who accepted the first dose of the
vaccine who go on to accept the second dose, becoming fully vaccinated}

\item{r2_p}{proportion of partially vaccinated individuals who receive
a second dose when returning to the clinic due to screening or illness}

\item{booster_uptake}{proportion of the formerly fully vaccinated, waned
population who accept a booster vaccination dose}

\item{n_erlang}{integer giving the number of transitions that need to be
made
through vaccine-protected strata until that protection has waned}
}
\value{
an array of the uptakes of same dimensions
}
\description{
Creates uptake mapping for the branching XPVWRH model where
individuals can move from unvaccinated (X) to vaccinated (V) or partially
vaccinated (P) as well as revaccinated from waned (W) to (R) and, and
partially vaccinated (P) to fully vaccianted (V). The former
reflects the specific indices which are chosen for assigning uptakes.
}
