% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uptake_maps.R
\name{create_uptake_map_xvwr}
\alias{create_uptake_map_xvwr}
\title{Creates uptake mapping array with dimensions n_group x n_vax x n_vax
and assigns the relevant primary uptake and booster uptake values defined
by the user.}
\usage{
create_uptake_map_xvwr(
  n_group,
  n_vax,
  primary_uptake,
  booster_uptake,
  idx,
  n_diag_rec = 1,
  screening_or_diagnosis
)
}
\arguments{
\item{n_group}{scalar indicating number of activity groups}

\item{n_vax}{scalar indicating number the number of stratum in the model}

\item{primary_uptake}{proportion of the unvaccinated population who accept
primary vaccination}

\item{booster_uptake}{proportion of the formerly fully vaccinated, waned
population who accept a booster vaccination dose}

\item{idx}{list containing indices of all X,  V, W, R strata and n_vax
through vaccine-protected strata until that protection has waned}

\item{n_diag_rec}{integer for the number of diagnosis history substrata}

\item{screening_or_diagnosis}{string indicating screening or diagnosis}
}
\value{
an array of the uptakes with dimensions n_group x n_vax x n_vax
}
\description{
Creates uptake mapping array with dimensions n_group x n_vax x n_vax
and assigns the relevant primary uptake and booster uptake values defined
by the user.
}
