% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_vax_map}
\alias{create_vax_map}
\title{Create mapping for movement between strata due to vaccination}
\usage{
create_vax_map(n_vax, v, i_u, i_v)
}
\arguments{
\item{n_vax}{Integer in (0, 5) denoting total number of strata}

\item{v}{a matrix of dimensions 2x2 (vod, vos) or a scalar (vbe) indicating
% of population vaccinated according to the strategy. For vod & vos each
value will apply to the eligible stratum separately}

\item{i_u}{indices of strata eligible for vaccination}

\item{i_v}{indices of strata being vaccinated}
}
\value{
an array of the mapping
}
\description{
Create mapping for movement between strata due to vaccination
}
