% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xpvwrh.R
\name{create_vax_map_branching}
\alias{create_vax_map_branching}
\title{Create mapping for movement between strata due to vaccination where
vaccination uptake splits off into two types (partial and full) from the
naive population (X)}
\usage{
create_vax_map_branching(
  n_vax,
  v,
  i_u,
  i_v,
  r1 = NULL,
  r1r2 = NULL,
  set_vbe = FALSE
)
}
\arguments{
\item{n_vax}{Integer denoting total number of strata}

\item{v}{0-1 vector of length two indicating whether activity group
should be offered vaccination.}

\item{i_u}{indices of strata eligible for vaccination}

\item{i_v}{indices of strata vaccinated and protected}
}
\value{
an array of the mapping
}
\description{
Create mapping for movement between strata due to vaccination where
vaccination uptake splits off into two types (partial and full) from the
naive population (X)
}
