% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_waning_map}
\alias{create_waning_map}
\title{Create mapping for movement between strata due to vaccine waning}
\usage{
create_waning_map(n_vax, i_w, z)
}
\arguments{
\item{n_vax}{Integer in (0, 5) denoting total number of strata}

\item{i_w}{Integer in (0, 5) denoting which stratum receives waned vaccinees}

\item{z}{Scalar denoting rate of waning}
}
\value{
an array of the mapping
}
\description{
Create mapping for movement between strata due to vaccine waning
}
