% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_trial.R
\name{create_waning_map_trial}
\alias{create_waning_map_trial}
\title{Create mapping for movement between strata due to vaccine waning
in a vaccine trial with erlang compartments}
\usage{
create_waning_map_trial(n_vax, i_v, i_w, z)
}
\arguments{
\item{n_vax}{Integer in (0, 5) denoting total number of strata}

\item{i_v}{indices of strata under vaccination protection}

\item{i_w}{indicies denoting which stratum receives waned vaccinees}

\item{z}{Scalar denoting rate of waning}
}
\value{
an array of the mapping
}
\description{
Create mapping for movement between strata due to vaccine waning
in a vaccine trial with erlang compartments
}
