% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualvax.R
\name{dualvax_initial}
\alias{dualvax_initial}
\title{Initial conditions for the dualvax model}
\usage{
dualvax_initial(contents)
}
\arguments{
\item{pars}{A parameter list created by [dualvax_parameters()]; from
this list we will use the `N0` `q`, `prev_Asl` and `prev_Ash`
elements.}
}
\value{
A list of initial conditions
}
\description{
Create initial conditions for the dualvax model
}
