% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualvax.R
\name{dualvax_params}
\alias{dualvax_params}
\title{Parameters for the dualvax model}
\usage{
dualvax_params(gono_params = NULL, init_params = NULL, vax_params = NULL)
}
\arguments{
\item{gono_params}{A dataframe of natural history parameters}

\item{init_params}{A list of starting conditions}

\item{vax_params}{A vector of vaccination params}
}
\value{
A list of inputs to the model many of which are fixed and
  represent data. These correspond largely to `user()` calls
  within the odin code, though some are also used in processing
  just before the model is run.
}
\description{
Parameters for the dualvax model
}
