% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualvax.R
\name{dualvax_restart}
\alias{dualvax_restart}
\title{Create parameters to start the dualvax model from equilibrium}
\usage{
dualvax_restart(n_vax, comps)
}
\arguments{
\item{n_vax}{an integer indicating the number of vaccine compartments}

\item{comps}{a named list of arrays, as output by the model
(including U, I, A, S, T)}
}
\value{
A list of initial conditions
}
\description{
Create parameters to start the dualvax model from equilibrium
}
