% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{extract_flows}
\alias{extract_flows}
\title{extract flows used for run_grid}
\usage{
extract_flows(y)
}
\arguments{
\item{y}{a transformed model run output}
}
\value{
cumulative and incident flows
}
\description{
extract flows used for run_grid
}
