% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{extract_flows_trial}
\alias{extract_flows_trial}
\title{extract flows for the XVW trial model}
\usage{
extract_flows_trial(y)
}
\arguments{
\item{y}{a transformed model run output}
}
\value{
cumulative and incident flows
}
\description{
extract flows for the XVW trial model
}
