% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{extract_flows_xpvwrh}
\alias{extract_flows_xpvwrh}
\title{extract flows used for run_grid when the branching  xpvwrh model
has been run}
\usage{
extract_flows_xpvwrh(y)
}
\arguments{
\item{y}{a transformed model run output}
}
\value{
cumulative and incident flows
}
\description{
extract flows used for run_grid when the branching  xpvwrh model
has been run
}
