% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xpvwrh.R
\name{gen_wane_vac}
\alias{gen_wane_vac}
\alias{gen_wane_vec}
\title{generates 'i_v' or 'i_w' vectors which contain indices of strata
in the model where people wane from or wane to to later generate
waning maps}
\usage{
gen_wane_vec(n_erlang, n_vax, i_p, direction)
}
\arguments{
\item{n_erlang}{integer giving the number of transitions that need to be
 made
through vaccine-protected strata until that protection has waned}

\item{n_vax}{integer denoting total number of strata}

\item{i_p}{vector containing indices of strata into which individuals are
vaccinated}

\item{direction}{string, either "from" or "to" which switches the action of
the function between generating i_w or i_v}
}
\value{
a character vector of length 3 containing indices of strata where
people wane to or from depending on n_vax
}
\description{
generates 'i_v' or 'i_w' vectors which contain indices of strata
in the model where people wane from or wane to to later generate
waning maps
}
