% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{gono_params}
\alias{gono_params}
\title{Posterior parameters of gonorrhoea natural history}
\usage{
gono_params(n = NULL)
}
\arguments{
\item{n}{an integer vector (or value) containing the indices of the required
parameter sets (1:982). If `n = NULL` the full parameter set is returned}
}
\value{
A data frame of parameters
}
\description{
Posterior parameters of gonorrhoea natural history
}
