% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gonovax_data}
\alias{gonovax_data}
\title{Get gonovax data}
\usage{
gonovax_data()
}
\value{
A data.frame containing the gonovax year (i.e. year after 2007),
the year, the total diagnoses in MSM, the total tests in MSM and the source
of the data, the proportion of symptomatic diagnoses
}
\description{
Get annual time series of GUMCAD diagnoses and testing figures, and annual
proportion symptomatic from GRASP
}
