% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{gonovax_year}
\alias{gonovax_year}
\title{Convert a year into the number of years after 2009}
\usage{
gonovax_year(year)
}
\arguments{
\item{year}{an integer year}
}
\value{
An integer, being the number of years after 2009
}
\description{
Convert a year into the number of years after 2009
}
\examples{
gonovax_year(2019)
gonovax_year(c(2018, 2019))
}
