% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{gonovax_year_as_year}
\alias{gonovax_year_as_year}
\title{Convert a gonovax year into calendar years}
\usage{
gonovax_year_as_year(gonovax_year)
}
\arguments{
\item{gonovax_year}{an integer}
}
\value{
An integer, being the calendar year
}
\description{
Convert a gonovax year into calendar years
}
\examples{
gonovax_year_as_year(3)
}
