% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gumcad_data}
\alias{gumcad_data}
\title{Get GUMCAD data}
\usage{
gumcad_data()
}
\value{
A data.frame containing the gonovax year (i.e. year after 2007),
the year, the total diagnoses in MSM, the total tests in MSM and the source
of the data
}
\description{
Get annual time series of GUMCAD diagnoses and testing figures
}
