% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{initial_params}
\alias{initial_params}
\title{Initial conditions for the model}
\usage{
initial_params(pars, n_vax = 1)
}
\arguments{
\item{pars}{A parameter list containing `N0`, `q`, `prev_Asl` and `prev_Ash`
elements.}

\item{n_vax}{an integer indicating the number of vaccine compartments}
}
\value{
A list of initial conditions
}
\description{
Create initial conditions for the model
}
