% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_repeated_xpvwr.R
\name{initial_params_repeated_xpvwr}
\alias{initial_params_repeated_xpvwr}
\title{Initial conditions for the model}
\usage{
initial_params_repeated_xpvwr(
  pars,
  hesgroups = 1,
  groupsize = rep(1, hesgroups)/hesgroups,
  coverage_p = rep(0, hesgroups),
  coverage_v = rep(0, hesgroups),
  t = FALSE,
  n_erlang = 1,
  n_diag_rec = 1
)
}
\arguments{
\item{pars}{A parameter list containing `N0`, `q`, `prev_Asl` and `prev_Ash`
elements.}

\item{hesgroups}{number of different vaccine hesitancy groups 
(with different vaccine sentiments)}

\item{coverage_p}{now ARRAY of partial (one-dose) vaccine coverage of 
the population already present (as a proportion)}

\item{coverage_v}{now ARRAY of two-dose vaccine coverage of 
the population already present (as a proportion)}

\item{t}{number of years, only use when using function outside of
run_onevax_xpvwrh() to generate initial conditions for tests}

\item{n_erlang}{integer giving the number of transitions that need to be
made through vaccine-protected strata until that protection has waned}

\item{n_diag_rec}{number of diagnosis history strata}

\item{coverage_x}{now ARRAY of size of hesitancy groups}
}
\value{
A list of initial conditions
}
\description{
Create initial conditions for the model
}
