% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xpvwrh.R
\name{initial_params_xpvwrh}
\alias{initial_params_xpvwrh}
\title{Initial conditions for the model}
\usage{
initial_params_xpvwrh(pars, coverage_p = 0, coverage_v = 0, hes = 0, t = FALSE)
}
\arguments{
\item{pars}{A parameter list containing `N0`, `q`, `prev_Asl` and `prev_Ash`
elements.}

\item{hes}{proportion of population vaccine hesitant}

\item{t}{number of years, only use when using function outside of
run_onevax_xpvwrh() to generate initial conditions for tests}
}
\value{
A list of initial conditions
}
\description{
Create initial conditions for the model
}
