% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc}
\alias{mcmc}
\title{Run a mcmc sampler}
\usage{
mcmc(pars, n_steps, compare = NULL, progress = FALSE, n_chains = 1)
}
\arguments{
\item{pars}{A [`mcmc_parameters`] object containing
information about parameters ( parameter ranges, priors, proposal kernel,
observation functions).}

\item{n_steps}{Number of MCMC steps to run}

\item{compare}{likelihood function to compare data to epidemic trajectory
should return a single value representing the log-likelihood. Default is
compare_basic()}

\item{progress}{Logical, indicating if a progress bar should be
displayed, using [`progress::progress_bar`].}

\item{n_chains}{Optional integer, indicating the number of chains
to run. If more than one then we run a series of chains and
merge them with [mcmc_combine()]. Chains are run in series,
with the same model.}
}
\value{
A `gonovax_mcmc` object containing `pars`
  (sampled parameters) and `probabilities` (log prior, log
  likelihood and log posterior values for these
  probabilities).
}
\description{
Run a mcmc sampler
}
\details{
This is a basic Metropolis-Hastings MCMC sampler.  The
`model` is run with a set of parameters to evaluate the
likelihood. A new set of parameters is proposed, and these
likelihoods are compared, jumping with probability equal to their
ratio. This is repeated for `n_steps` proposals.

This function is adapted from the `pmcmc` in the mcstate package
https://github.com/mrc-ide/mcstate
}
