% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-tools.R
\name{mcmc_combine}
\alias{mcmc_combine}
\title{Combine mcmc samples}
\usage{
mcmc_combine(..., samples = list(...))
}
\arguments{
\item{...}{Arguments representing [mcmc()] sample, i.e.,
`gonovax_mcmc` objects. Alternatively, pass a list as the
argument `samples`. Names are ignored.}

\item{samples}{A list of `gonovax_mcmc` objects. This is often
more convenient for programming against than `...`}
}
\description{
all functions in mcmc-tools.R are adapted from pmcmc-tools.R functions in
github.com/mrc-ide/mcstate
Combine multiple [mcmc()] samples into one object
}
