% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-tools.R
\name{mcmc_thin}
\alias{mcmc_thin}
\alias{mcmc_sample}
\title{Thin a mcmc chain}
\usage{
mcmc_thin(object, burnin = NULL, thin = NULL)

mcmc_sample(object, n_sample, burnin = NULL)
}
\arguments{
\item{object}{Results of running [mcmc()]}

\item{burnin}{Optional integer number of iterations to discard as
"burn-in". If given then samples `1:burnin` will be
excluded from your results. Remember that the first sample
represents the starting point of the chain. It is an error if
this is not a positive integer or is greater than or equal to
the number of samples (i.e., there must be at least one sample
remaining after discarding burnin).}

\item{thin}{Optional integer thinning factor. If given, then every
`thin`'th sample is retained (e.g., if `thin` is 10
then we keep samples 1, 11, 21, ...).}

\item{n_sample}{The number of samples to draw from `object` *with
replacement*. This means that `n_sample` can be larger than the
total number of samples taken (though it probably should not)}
}
\description{
Thin results of running [mcmc()].`mcmc_thin` takes every `thin`'th sample,
 while `mcmc_sample` randomly selects a total of `n_sample` samples.
}
