% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{model_params}
\alias{model_params}
\title{Parameters for the dualvax model}
\usage{
model_params(
  gono_params = NULL,
  demographic_params = NULL,
  init_params = NULL,
  vax_params = NULL
)
}
\arguments{
\item{gono_params}{A dataframe of natural history parameters}

\item{demographic_params}{A dataframe of demographic parameters}

\item{init_params}{A list of starting conditions}

\item{vax_params}{A vector of vaccination params}
}
\value{
A list of inputs to the model many of which are fixed and
  represent data. These correspond largely to `user()` calls
  within the odin code, though some are also used in processing
  just before the model is run.
}
\description{
Parameters for the dualvax model
}
