% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_trial.R
\name{model_params_trial}
\alias{model_params_trial}
\title{Parameters for the vaccination trial model}
\usage{
model_params_trial(
  gono_params_trial = NULL,
  demographic_params_trial = NULL,
  initial_params_trial = NULL,
  vax_params = NULL,
  p_v = 0,
  n_erlang = 1
)
}
\arguments{
\item{gono_params_trial}{A dataframe of natural history parameters}

\item{demographic_params_trial}{A dataframe of demographic parameters}

\item{initial_params_trial}{A list of starting conditions}

\item{vax_params}{A vector of vaccination params}

\item{p_v}{A scalar indicating the percentage of the trial cohort that
is vaccinated}
}
\value{
A list of inputs to the model many of which are fixed and
  represent data. These correspond largely to `user()` calls
  within the odin code, though some are also used in processing
  just before the model is run.
}
\description{
Parameters for the vaccination trial model
}
