% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{model_params_xpvwrh}
\alias{model_params_xpvwrh}
\title{Parameters for the dualvax model}
\usage{
model_params_xpvwrh(
  gono_params = NULL,
  demographic_params = NULL,
  init_params = NULL,
  vax_params = NULL,
  n_erlang = 1,
  n_diag_rec = 1
)
}
\arguments{
\item{gono_params}{A dataframe of natural history parameters}

\item{demographic_params}{A dataframe of demographic parameters}

\item{init_params}{A list of starting conditions}

\item{vax_params}{A vector of vaccination params}

\item{n_erlang}{integer giving the number of transitions that need to be
made through vaccine-protected strata until that protection has waned}

\item{n_diag_rec}{integer for the number of diagnosis history substrata}
}
\value{
A list of inputs to the model many of which are fixed and
  represent data. These correspond largely to `user()` calls
  within the odin code, though some are also used in processing
  just before the model is run.
}
\description{
Parameters for the dualvax model
}
