% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{never_diag_hist_id}
\alias{never_diag_hist_id}
\title{For post-processing of the XVW stochastic trial model, generates
 indices of the strata in V in which individuals have not yet been diagnosed
 including erlang > 1}
\usage{
never_diag_hist_id(vec, n_erlang, n_diag_rec)
}
\arguments{
\item{vec}{a vector of indices for 'V' generated by 
stratum_index_xvw_trial()}

\item{n_erlang}{integer giving the number of transitions that need to be
made to wane entirely}

\item{n_diag_rec}{integer giving the number of each X, V(erlang), and W
stratum, allowing tracking of diagnosis history.}
}
\value{
vector of indices matching vaccinated strata in which diagnosis 
has not yet occurred
}
\description{
For post-processing of the XVW stochastic trial model, generates
 indices of the strata in V in which individuals have not yet been diagnosed
 including erlang > 1
}
