% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novax.R
\name{novax_baseline}
\alias{novax_baseline}
\title{Equilibrium state of compartmental model with no vaccination}
\usage{
novax_baseline(n = NULL, t)
}
\arguments{
\item{n}{an integer vector (or value) containing the indices of
corresponding parameter set (1:982). If `n = NULL` the equilibrium positions
for the full parameter set are returned}

\item{t}{an integer between 1 and 10 of the year to return}
}
\value{
A list containing incid and cum_incid at time t for parameter sets n
}
\description{
Equilibrium state of compartmental model with no vaccination
}
