% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novax.R
\name{novax_equilib}
\alias{novax_equilib}
\title{Equilibrium state of compartmental model with no vaccination}
\usage{
novax_equilib(n = NULL)
}
\arguments{
\item{n}{an integer vector (or value) containing the indices of
corresponding parameter set (1:982). If `n = NULL` the equilibrium positions
for the full parameter set are returned}
}
\value{
A list of transformed model outputs
}
\description{
Equilibrium state of compartmental model with no vaccination
}
