% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualvax.R
\name{Parameters for the "[dualvax]" model.}
\alias{Parameters for the "[dualvax]" model.}
\alias{dualvax_parameters}
\title{Parameters for the dulavax model}
\usage{
dualvax_parameters(gono_params = NULL, init_params = NULL, vax_params = NULL)
}
\arguments{
\item{gono_params}{A vector of natural history parameters}

\item{vax_params}{A vector of vaccination params}
}
\value{
A list of inputs to the model many of which are fixed and
  represent data. These correspond largely to `user()` calls
  within the odin code, though some are also used in processing
  just before the model is run.
}
\description{
Parameters for the dulavax model
}
