% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{restart_params}
\alias{restart_params}
\title{Create initial conditions to start the model from the end of a run}
\usage{
restart_params(y, n_vax = NULL)
}
\arguments{
\item{y}{a transformed model run output}

\item{n_vax}{an integer indicating the number of vaccine compartments,
consistent with the input}
}
\value{
A list of initial conditions to restart a model with n_vax
vaccination levels
}
\description{
Create initial conditions based on previous model run
}
