% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run}
\alias{run}
\title{Run odin model of gonorrhoea with or without vaccination}
\usage{
run(
  tt,
  gono_params,
  demographic_params = NULL,
  init_params = NULL,
  vax_params = NULL,
  n_diag_rec = 1,
  transform = TRUE
)
}
\arguments{
\item{tt}{a numeric vector of times at which the model state is output}

\item{gono_params}{a data frame of parameters}

\item{demographic_params}{= NULL}

\item{init_params}{= NULL}

\item{vax_params}{= NULL}

\item{n_diag_rec}{integer for the number of diagnosis history substrata}

\item{transform}{= TRUE}
}
\description{
Run odin model of gonorrhoea with or without vaccination
}
