% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax.R
\name{run_grid_onevax}
\alias{run_grid_onevax}
\title{run model from equilibrium with single vaccine at the input 
efficacy / duration grid locations for n parameter sets}
\usage{
run_grid_onevax(n, t, eff, dur, ve = 0, vd = 0, vs = 0)
}
\arguments{
\item{n}{a value denoting the number of parameter sets to run}

\item{t}{an integer number of years at which impact is to be assessed}

\item{eff}{numeric vector (between 0-1) of efficacy values of the vaccine}

\item{dur}{numeric vector of duration in years of the vaccine}

\item{ve}{single numeric indicating % of population vaccinated before entry
(between 0-1)}

\item{vd}{single numeric indicating % of population vaccinated on diagnosis
(between 0-1)}

\item{vs}{single numeric indicating % of population vaccinated on screening
(between 0-1)}
}
\value{
A `gonovax_grid` object
}
\description{
run model from equilibrium with single vaccine at the input 
efficacy / duration grid locations for n parameter sets
}
