% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novax.R
\name{run_novax}
\alias{run_novax}
\title{run model without vaccination for input parameter sets, either from
initialisation or from equilibrium}
\usage{
run_novax(n = NULL, tt, equilib = FALSE)
}
\arguments{
\item{n}{an integer vector (or value) containing the indices of
corresponding parameter set (1:982). If `n = NULL` the equilibrium positions
for the full parameter set are returned}

\item{tt}{a numeric vector of times at which the model state is output}

\item{equilib}{a logical indicating whether to run from equilibrium, default
is `FALSE`, i.e. run from initial conditions}
}
\value{
A list of transformed model outputs
}
\description{
run model without vaccination for input parameter sets, either from
initialisation or from equilibrium
}
