% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax.R
\name{run_onevax}
\alias{run_onevax}
\title{run model with single vaccine for input parameter sets, either from
initialisation or from equilibrium}
\usage{
run_onevax(n = NULL, tt, eff, dur, ve = 0, vd = 0, vs = 0, equilib = FALSE)
}
\arguments{
\item{n}{an integer vector (or value) containing the indices of
corresponding parameter set (1:982). If `n = NULL` the equilibrium positions
for the full parameter set are returned}

\item{tt}{a numeric vector of times at which the model state is output}

\item{eff}{single numeric indicating efficacy of the vaccine (between 0-1)}

\item{dur}{single numeric indicating duration of the vaccine (in years)}

\item{ve}{single numeric indicating % of population vaccinated before entry
(between 0-1)}

\item{vd}{single numeric indicating % of population vaccinated on diagnosis
(between 0-1)}

\item{vs}{single numeric indicating % of population vaccinated on screening
(between 0-1)}

\item{equilib}{a logical indicating whether to run from equilibrium, default
is `FALSE`, i.e. run from initial conditions}
}
\value{
A list of transformed model outputs
}
\description{
run model with single vaccine for input parameter sets, either from
initialisation or from equilibrium
}
