% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax.R
\name{run_onevax_user}
\alias{run_onevax_user}
\title{run model with vaccination for input parameter sets from
initialisation}
\usage{
run_onevax_user(gono_params, tt)
}
\arguments{
\item{gono_params}{a set of gono_params, eg a mcmc sample}

\item{tt}{a numeric vector of times at which the model state is output}
}
\value{
A list of transformed model outputs
}
\description{
run model with vaccination for input parameter sets from
initialisation
}
