% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvw.R
\name{run_onevax_xvw}
\alias{run_onevax_xvw}
\title{Run model with single vaccine for input parameter sets, either from
initialisation or from equilibrium, those with waned vaccines are not
eligible for revaccination.}
\usage{
run_onevax_xvw(
  tt,
  gono_params,
  init_params = NULL,
  dur = 1000,
  vea = 0,
  vei = 0,
  ved = 0,
  ves = 0,
  vbe = 0,
  uptake = 0,
  strategy = "VbE",
  coverage = 0,
  t_stop = 99
)
}
\arguments{
\item{tt}{a numeric vector of times at which the model state is output}

\item{gono_params}{list of gono params}

\item{init_params}{= NULL}

\item{dur}{scalar or numeric vector with same length as `gono_params`
giving duration of the vaccine (in years)}

\item{vea}{scalar or numeric vector with same length as `gono_params` giving
efficacy of the vaccine against acquisition (between 0-1)}

\item{vei}{scalar or numeric vector with same length as `gono_params` giving
efficacy of the vaccine against infectiousness (between 0-1)}

\item{ved}{scalar or numeric vector with same length as `gono_params` giving
efficacy of the vaccine against duration (between 0-1)}

\item{ves}{scalar or numeric vector with same length as `gono_params` giving
efficacy of the vaccine against symptoms (between 0-1)}

\item{vbe}{scalar indicating pc of population vaccinated before entry
(between 0-1)}

\item{uptake}{scalar or numeric vector with same length as `gono_params`
giving pc of population vaccinated as part of strategy}

\item{strategy}{single character string in "VbE", "VoD", "VoD(H)",
"VoA", "VoA(H)", "VoD(L)+VoA(H)"}

\item{coverage}{scalar giving initial coverage of vaccination}

\item{t_stop}{time at which vaccination should stop (years)}
}
\description{
Run model with single vaccine for input parameter sets, either from
initialisation or from equilibrium, those with waned vaccines are not
eligible for revaccination.
}
