% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvw_trial.R
\name{run_onevax_xvw_trial}
\alias{run_onevax_xvw_trial}
\title{Run model with single vaccine for input parameter sets, either from
initialisation or from equilibrium, those with waned vaccines are not
eligible for re-vaccination.}
\usage{
run_onevax_xvw_trial(
  tt,
  gono_params,
  initial_params_trial = NULL,
  dur = 1000,
  vea = 0,
  vei = 0,
  ved = 0,
  ves = 0,
  p_v = 0.5,
  n_erlang = 1,
  stochastic = FALSE,
  N = 6e+05,
  n_diag_rec = 1,
  asymp_recorded = TRUE
)
}
\arguments{
\item{tt}{a numeric vector of times at which the model state is output in
years}

\item{gono_params}{list of gono params for a vaccination trial}

\item{initial_params_trial}{list of initial conditions for model trial. Set
default as NULL.}

\item{dur}{scalar or numeric vector with same length as `gono_params`
giving duration of the vaccine (in years)}

\item{vea}{scalar or numeric vector with same length as `gono_params` giving
efficacy of the vaccine against acquisition (between 0-1)}

\item{vei}{scalar or numeric vector with same length as `gono_params` giving
efficacy of the vaccine against infectiousness (between 0-1)}

\item{ved}{scalar or numeric vector with same length as `gono_params` giving
efficacy of the vaccine against duration (between 0-1)}

\item{ves}{scalar or numeric vector with same length as `gono_params` giving
efficacy of the vaccine against symptoms (between 0-1)}

\item{p_v}{scalar giving proportion of the trial cohort vaccinated, default
is 0.5.}

\item{n_erlang}{integer giving the number of transitions that need to be
made}

\item{stochastic}{logical indicating if the run should be made with the
default deterministic trial model in continuous time or stochastic trial
model in discrete time}

\item{N}{integer to assign the total number of individuals in the trial
(split equally across the two arms)}

\item{n_diag_rec}{integer giving the number of each X, V(erlang), and W
stratum, allowing tracking of diagnosis history. e.g for a n_diag_rec = 2
and erlang = 1, there will be X.I, X.II, V1.I, V1.II, W.I, W.II strata.
Where '.I' corresponds to never-diagnosed individuals and '.II' is for
individuals diagnosed at least once.}

\item{asymp_recorded}{logical indicating if the trial screens for and
records asymptomatic diagnosis. If FALSE, asymptomatic infected individuals
undergoing treatment do not move diagnosis history stratum}
}
\description{
Run model with single vaccine for input parameter sets, either from
initialisation or from equilibrium, those with waned vaccines are not
eligible for re-vaccination.
}
