% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvwr.R
\name{run_onevax_xvwr}
\alias{run_onevax_xvwr}
\title{run model with single vaccine for input parameter sets, either from
initialisation or from equilibrium, those with waned vaccines are eligible
for revaccination (R), and return to the R stratum}
\usage{
run_onevax_xvwr(
  tt,
  gono_params,
  init_params = NULL,
  dur,
  vea = 0,
  vei = 0,
  ved = 0,
  ves = 0,
  dur_revax = dur,
  vea_revax = vea,
  vei_revax = vei,
  ved_revax = ved,
  ves_revax = ves,
  vbe = 0,
  uptake = 0,
  strategy = "VbE",
  t_stop = 99
)
}
\arguments{
\item{tt}{a numeric vector of times at which the model state is output}

\item{gono_params}{list of gono params}

\item{init_params}{= NULL}

\item{dur}{scalar or numeric vector with same length as `gono_params`
giving duration of the vaccine (in years)}

\item{vea}{scalar or numeric vector with same length as `gono_params` giving
efficacy of the vaccine against acquisition (between 0-1)}

\item{vei}{scalar or numeric vector with same length as `gono_params` giving
efficacy of the vaccine against infectiousness (between 0-1)}

\item{ved}{scalar or numeric vector with same length as `gono_params` giving
efficacy of the vaccine against duration (between 0-1)}

\item{ves}{scalar or numeric vector with same length as `gono_params` giving
efficacy of the vaccine against symptoms (between 0-1)}

\item{dur_revax}{scalar or numeric vector with same length as `gono_params`
giving duration of protection for revaccination, default to same as primary}

\item{vea_revax}{scalar or numeric vector with same length as `gono_params`
giving efficacy of revaccination against acquisition, default to same as
primary}

\item{vei_revax}{scalar or numeric vector with same length as `gono_params`
giving efficacy of revaccination against infectiousness, default to same as
primary}

\item{ved_revax}{scalar or numeric vector with same length as `gono_params`
giving efficacy of revaccination against duration of infection, default to
same as primary}

\item{ves_revax}{scalar or numeric vector with same length as `gono_params`
giving efficacy of revaccination against symptoms, default to same as
primary}

\item{vbe}{scalar indicating pc of population vaccinated before entry
(between 0-1)}

\item{uptake}{scalar or numeric vector with same length as `gono_params`
giving pc of population vaccinated as part of strategy}

\item{strategy}{single character string in "VbE", "VoD", "VoD(H)",
"VoA", "VoA(H)", "VoD(L)+VoA(H)"}

\item{t_stop}{time at which vaccination should stop (years)}
}
\value{
A list of transformed model outputs
}
\description{
run model with single vaccine for input parameter sets, either from
initialisation or from equilibrium, those with waned vaccines are eligible
for revaccination (R), and return to the R stratum
}
