% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trial.R
\name{run_trial}
\alias{run_trial}
\title{Run odin model of gonorrhoea vaccine trial with or without
vaccination}
\usage{
run_trial(
  tt,
  gono_params,
  init_params = NULL,
  vax_params = NULL,
  transform = TRUE
)
}
\arguments{
\item{tt}{a numeric vector of times at which the model state is output}

\item{gono_params}{a data frame of parameters}

\item{init_params}{= NULL}

\item{vax_params}{= NULL}

\item{transform}{= TRUE}
}
\description{
Run odin model of gonorrhoea vaccine trial with or without
vaccination
}
