% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xpvwrh.R
\name{set_protection}
\alias{set_protection}
\title{generates vector which tells the model which strata are under
vaccine protection and what the value of protection for that strata is}
\usage{
set_protection(i_v, n_erlang, n_vax, ve_p, ve, ve_revax)
}
\arguments{
\item{i_v}{indices of strata receiving protection through vaccination}

\item{n_erlang}{scalar giving the number of transitions that need to be made
through vaccine-protected strata until that protection has waned}

\item{n_vax}{integer denoting total number of strata}

\item{ve_p}{scalar 0-1 with degree of partial primary protection of the P(N)
strata, can take vea_p, vei_p, ves_p, ved_p}

\item{ve}{scalar 0-1 with degree of full primary protection of the V(N)
strata, can take vea, vei, ves, ved}

\item{ve_revax}{scalar 0-1 with degree of re-vaccinated protection of the
R(N) strata, can take vea_revax, vei_revax, ves_revax, ved_revax}
}
\value{
vector of length n_vax with zeros corresponding to the indices of
strata with no protection, and the supplied degree of partial, full, and 
boosted protection corresponding to the indices of strata with partial,
full and boosted vaccination status
}
\description{
generates vector which tells the model which strata are under
vaccine protection and what the value of protection for that strata is
}
