% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{set_strategy}
\alias{set_strategy}
\title{Translate each named vaccine strategy into a format interpretable by
`create_vax_map`}
\usage{
set_strategy(strategy, include_vbe = FALSE)
}
\arguments{
\item{strategy}{single character string in "VbE", "VoD", "VoD(H)",
"VoA", "VoA(H)", "VoD(L)+VoA(H)", "VoS" describing who is offered
vaccination}

\item{include_vbe}{single logical indicating whether vaccination before
entry is offered.}
}
\value{
list with entries `vod`, `vos` and `vbe` containing 0-1 vectors of
length two indicating whether group L and/or H are offered vaccination via
each potential vaccination route (i.e. screeing, diagnosis, entry)
}
\description{
Translate each named vaccine strategy into a format interpretable by
`create_vax_map`
}
