% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratum_indexes.R
\name{stratum_index_repeated_xpvwr}
\alias{stratum_index_repeated_xpvwr}
\title{Generate the indices of all xpvwr strata}
\usage{
stratum_index_repeated_xpvwr(
  n_erlang = 1,
  n_diag_rec = 1,
  hesgroups = 1,
  strategy = NULL
)
}
\arguments{
\item{n_erlang}{integer giving the number of transitions that need to be
made through vaccine-protected strata until that protection has waned}

\item{n_diag_rec}{integer for the number of diagnosis history substrata}

\item{hesgroups}{integer for the number of hesitancy groups}

\item{strategy}{string of vaccination strategy being considered}
}
\value{
A list of strata with their indicies
}
\description{
Generate the indices of all xpvwr strata
}
