% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvwrh.R
\name{stratum_index_xvwrh}
\alias{stratum_index_xvwrh}
\title{Generate the indices of all xvwrh strata}
\usage{
stratum_index_xvwrh()
}
\value{
A list of strata with their indicies
}
\description{
Generate the indices of all xvwrh strata
}
