% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{transform}
\alias{transform}
\title{Transform fitted parameters into gonovax params}
\usage{
transform(pars, fix_par_t = TRUE)
}
\arguments{
\item{pars}{list of fitted parameters}

\item{fix_par_t}{logical indicating whether parameters with inferred trends
are held at their 2020 values after that date.}
}
\value{
A list of parameters for use in the model
}
\description{
Transform fitted parameters into gonovax params
}
