% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{transform_fixed}
\alias{transform_fixed}
\title{Transform fitted parameters into non-time-varying gonovax params}
\usage{
transform_fixed(pars)
}
\arguments{
\item{pars}{list of fitted parameters}
}
\value{
A list of parameters for use in the model
}
\description{
Transform fitted parameters into non-time-varying gonovax params
}
