% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novax.R
\name{vax_params0_repeat}
\alias{vax_params0_repeat}
\title{create vaccination parameters for use in novax model (null)}
\usage{
vax_params0_repeat(n_diag_rec = 1, years_history = 1, hesgroups = 1)
}
\arguments{
\item{n_diag_rec}{integer giving the number of each X, V(erlang), and W}

\item{years_history}{number of years that diagnosis history is recorded for
stratum, allowing tracking of diagnosis history. e.g for a n_diag_rec = 2
and erlang = 1, there will be X.I, X.II, V1.I, V1.II, W.I, W.II strata.
Where '.I' corresponds to never-diagnosed individuals and '.II' is for
individuals diagnosed at least once.}
}
\value{
A list parameters in the model input format
}
\description{
create vaccination parameters for use in novax model (null)
}
