% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xpvwrh.R
\name{vax_params_xpvwrh}
\alias{vax_params_xpvwrh}
\title{create vaccination parameters for use in onevax_xpvwrh model}
\usage{
vax_params_xpvwrh(
  vea = 0,
  vei = 0,
  ved = 0,
  ves = 0,
  vea_revax = vea,
  vei_revax = vei,
  ved_revax = ved,
  ves_revax = ves,
  vea_p = vea,
  vei_p = vei,
  ved_p = ved,
  ves_p = ves,
  dur_v = 1000,
  dur_p = dur_v,
  dur_revax = dur_v,
  r1 = 0,
  r2 = 0,
  r2_p = 0,
  booster_uptake = r1 * r2,
  strategy = NULL,
  vbe = 0,
  t_stop = 99,
  hes = 0,
  n_erlang = 1,
  n_diag_rec = 1
)
}
\arguments{
\item{vea}{scalar indicating efficacy of the vaccine against acquisition
(between 0-1)}

\item{vei}{scalar indicating efficacy of the vaccine against infectiousness
(between 0-1)}

\item{ved}{scalar indicating efficacy of the vaccine against duration
(between 0-1)}

\item{ves}{scalar indicating efficacy of the vaccine against symptoms
(between 0-1)}

\item{vea_revax}{scalar indicating efficacy of revaccination against
acquisition (between 0-1)}

\item{vei_revax}{scalar indicating efficacy of revaccination against
infectiousness (between 0-1)}

\item{ved_revax}{scalar indicating efficacy of revaccination against
duration of infection (between 0-1)}

\item{ves_revax}{scalar indicating efficacy of revaccination against
symptoms (between 0-1)}

\item{vea_p}{scalar indicating efficacy of partial vaccination against
acquisition (between 0-1)}

\item{vei_p}{scalar indicating efficacy of partial vaccination against
infectiousness (between 0-1)}

\item{ved_p}{scalar indicating efficacy of partial vaccination against
duration (between 0-1)}

\item{ves_p}{scalar indicating efficacy of partial vaccination against
symptoms (between 0-1)}

\item{dur_v}{duration of time spent in V stratum after completing a round of
primary vaccination (fully vaccinated, accepting first and second dose)}

\item{dur_p}{duration of time spent in the P stratum, partially vaccinated
(accepting only the first dose)}

\item{dur_revax}{duration of protection for revaccination,
default to same as primary}

\item{r1}{proportion of population offered vaccine only accepting the first
dose, becoming partially vaccinated}

\item{r2}{proportion of the population who accepted the first dose of the
vaccine who go on to accept the second dose, becoming fully vaccinated}

\item{r2_p}{proportion of partially vaccinated individuals who receive
a second dose when returning to the clinic due to screening or illness}

\item{booster_uptake}{scalar or numeric vector with same length as
'gono_params' giving proportion of population undertaking booster
vaccination after primary vaccination protection has waned
@param n_diag_rec integer for the number of diagnosis history substrata}

\item{strategy}{single character string in "VoD", "VoD(H)",
"VoA", "VoA(H)", "VoD(L)+VoA(H)". Defaults to NULL i.e. no vaccination}

\item{vbe}{scalar indicating pc of population vaccinated before entry
(between 0-1)}

\item{t_stop}{time at which vaccination should stop (years)}

\item{hes}{proportion of population vaccine hesitant}

\item{n_erlang}{integer giving the number of transitions that need to be
made through vaccine-protected strata until that protection has waned}

\item{n_diag_rec}{number of diagnosis history strata}
}
\value{
A list parameters in the model input format
}
\description{
create vaccination parameters for use in onevax_xpvwrh model
}
