% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvw.R
\name{vax_params_xvw}
\alias{vax_params_xvw}
\title{Create vaccination parameters for use in onevax_xvw model}
\usage{
vax_params_xvw(
  vea = 0,
  vei = 0,
  ved = 0,
  ves = 0,
  dur = 1000,
  uptake = 0,
  strategy = NULL,
  vbe = 0,
  t_stop = 99
)
}
\arguments{
\item{vea}{scalar indicating efficacy of the vaccine against acquisition
(between 0-1)}

\item{vei}{scalar indicating efficacy of the vaccine against infectiousness
(between 0-1)}

\item{ved}{scalar indicating efficacy of the vaccine against duration
(between 0-1)}

\item{ves}{scalar indicating efficacy of the vaccine against symptoms
(between 0-1)}

\item{dur}{scalar indicating duration of the vaccine (in years)}

\item{uptake}{scalar indicating pc of those offered who accept vaccination}

\item{strategy}{single character string in "VoD", "VoD(H)",
"VoA", "VoA(H)", "VoD(L)+VoA(H)". Defaults to NULL i.e. no vaccination}

\item{vbe}{scalar indicating pc of population vaccinated before entry
(between 0-1)}

\item{t_stop}{time at which vaccination should stop (years)}
}
\value{
A list parameters in the model input format
}
\description{
Create vaccination parameters for use in onevax_xvw model
}
