% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvwr.R
\name{vax_params_xvwr}
\alias{vax_params_xvwr}
\title{create vaccination parameters for use in onevax_xvwr model}
\usage{
vax_params_xvwr(
  eff = 0,
  eff_revax = eff,
  dur = 1000,
  dur_revax = dur,
  uptake = 0,
  strategy = "VbE",
  ve = 0,
  t_stop = 99
)
}
\arguments{
\item{eff}{scalar indicating efficacy of the vaccine (between 0-1)}

\item{eff_revax}{efficacy of revaccination, default to same as primary}

\item{dur}{scalar indicating duration of the vaccine (in years)}

\item{dur_revax}{duration of protection for revaccination,
default to same as primary}

\item{uptake}{scalar indicating pc of population vaccinated as part
of strategy}

\item{strategy}{single character string in "VbE", "VoD(all)", "VoD(H)",
"VoA(all)", "VoA(H)", "VoD(L)+VoA(H)"}

\item{ve}{scalar indicating pc of population vaccinated before entry
(between 0-1)}

\item{t_stop}{time at which vaccination should stop (years)}
}
\value{
A list parameters in the model input format
}
\description{
create vaccination parameters for use in onevax_xvwr model
}
