% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{aggregate}
\alias{aggregate}
\title{aggregate model runs by vaccination group x vaccine strata}
\usage{
aggregate(x, what, as_incid = FALSE, f = identity, ...)
}
\arguments{
\item{x}{a transformed model run output}

\item{what}{a character string of the cumulative trajectory to transform}

\item{as_incid}{logical specifying whether to convert to incidence}

\item{f}{function to apply across parameter sets (eg. mean / median)}
}
\value{
a transformed time series / array
}
\description{
aggregate model runs by vaccination group x vaccine strata
}
