% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{compare}
\alias{compare}
\title{Calculate the log likelihood of the data given the parameters}
\usage{
compare(pars)
}
\arguments{
\item{pars}{A named vector of parameters}
}
\value{
a single log likelihood
}
\description{
Calculate the log likelihood of the data given the parameters
}
