% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xpvwrh.R
\name{create_uptake_map}
\alias{create_uptake_map}
\title{Creats uptake mapping for the branching XPVWRH model where
individuals can move from unvaccinated (X) to vaccinated (V) or partially
vaccinated (P) as well  as revaccinated from waned (W) to (R). The former
reflects the specific indices which are chosen for assigning uptakes.}
\usage{
create_uptake_map(array, r1, r2, booster_uptake)
}
\arguments{
\item{array}{a vaccine map array of dimensions n_group by n_vax by n_vax
generated through create_vax_map_branching()}

\item{r1}{proportion of population offered vaccine only accepting the first
dose, becoming partially vaccinated}

\item{r2}{proportion of the population who accepted the first dose of the
vaccine who go on to accept the second dose, becoming fully vaccinated}

\item{booster_uptake}{proportion of the formerly fully vaccinated, waned
population who accept a booster vaccination dose}
}
\value{
an array of the uptakes of same dimensions
}
\description{
Creats uptake mapping for the branching XPVWRH model where
individuals can move from unvaccinated (X) to vaccinated (V) or partially
vaccinated (P) as well  as revaccinated from waned (W) to (R). The former
reflects the specific indices which are chosen for assigning uptakes.
}
