% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_vax_map}
\alias{create_vax_map}
\title{Create mapping for movement between strata due to vaccination}
\usage{
create_vax_map(n_vax, v)
}
\arguments{
\item{n_vax}{Integer in (0, 5) denoting total number of strata}

\item{v}{numeric indicating % of population vaccinated on screening a scalar
will apply to both activity groups, a vector of length 2 will apply to each
activity group separately}
}
\value{
an array of the mapping
}
\description{
Create mapping for movement between strata due to vaccination
}
