% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_vax_map}
\alias{create_vax_map}
\title{Create mapping for movement between strata due to vaccination}
\usage{
create_vax_map(n_vax, v, i_u, i_v)
}
\arguments{
\item{n_vax}{Integer denoting total number of strata}

\item{v}{0-1 vector of length two indicating whether activity group
should be offered vaccination.}

\item{i_u}{indices of strata eligible for vaccination}

\item{i_v}{indices of strata being vaccinated}
}
\value{
an array of the mapping
}
\description{
Create mapping for movement between strata due to vaccination
}
