% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xpvwrh.R
\name{create_vax_map_branching}
\alias{create_vax_map_branching}
\title{Create mapping for movement between strata due to vaccination where
vaccination uptake splits off into two types (partial and full) from the
naive population (X). Different to create_vax_map as this function
specifically maps vbe to V (3) than P(2)}
\usage{
create_vax_map_branching(n_vax, v, i_u, i_v, set_vbe = FALSE, n_erlang = 1)
}
\arguments{
\item{n_vax}{Integer denoting total number of strata}

\item{v}{0-1 vector of length two indicating whether activity group
should be offered vaccination.}

\item{i_u}{indices of strata eligible for vaccination}

\item{i_v}{indices of strata vaccinated and protected}

\item{set_vbe}{Boolean which indicates that vaccination is occurring at some
level of uptake upon entering the model}

\item{n_erlang}{scalar giving the number of transitions that need to be made
through vaccine-protected strata until that protection has waned}
}
\value{
an array of the mapping
}
\description{
Create mapping for movement between strata due to vaccination where
vaccination uptake splits off into two types (partial and full) from the
naive population (X). Different to create_vax_map as this function
specifically maps vbe to V (3) than P(2)
}
