% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualvax.R
\name{dualvax_restart}
\alias{dualvax_restart}
\title{Create parameters to start the dualvax model from equilibrium}
\usage{
dualvax_restart(y, n_vax = NULL)
}
\arguments{
\item{y}{a transformed dualvax model run output}

\item{n_vax}{an integer indicating the number of vaccine compartments,
consistent with the input}
}
\value{
A list of initial conditions to restart a dualvax model with n_vax
vaccination levels
}
\description{
Create parameters to start the dualvax model from equilibrium
}
