% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{extract_flows_trial}
\alias{extract_flows_trial}
\title{extract flows for the XVW trial model}
\usage{
extract_flows_trial(y, idx)
}
\arguments{
\item{y}{a transformed model run output}

\item{idx}{list generated by stratum_index_xvw_trial() showing which
number strata belongs to which stratum type X, V or W}
}
\value{
cumulative and incident flows
}
\description{
extract flows for the XVW trial model
}
