% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_grid.R
\name{format_grid}
\alias{format_grid}
\title{format grid for heatmap plotting}
\usage{
format_grid(grid, disc_rate = 0, f = mean)
}
\arguments{
\item{grid}{a `gonovax_grid` object}

\item{disc_rate}{annual discount rate for cost-effectiveness calc,
e.g. 0.035 = 3.5pc, default is 0}

\item{f}{the function that should be used to summarise the runs}
}
\value{
a dataframe with columns denoting:
eff: efficacy of vaccine (%)
dur: duration of vaccine (years)
red_incid: Reduction in incidence after t years,
inc_vaccinated: Additional courses of vaccine over t years,
tot_red_incid: Infections averted over t years,
cost_eff: Courses of vaccine per infection averted (B / C),
discounted at `disc_rate`
tot_red_diag_a: Reduction in asymptomatic diagnoses over t years
tot_red_diag_s: Reduction in symptomatic diagnoses over t years
tot_inc_screened: Increase in screening over t years
}
\description{
format grid for heatmap plotting
}
