% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_grid.R
\name{format_grid}
\alias{format_grid}
\title{format grid for heatmap plotting}
\usage{
format_grid(grid)
}
\arguments{
\item{grid}{a `gonovax_grid` object}
}
\value{
a dataframe with columns denoting:
eff: efficacy of vaccine (%)
dur: duration of vaccine (years)
a: Reduction in incidence after t years,
b: Courses of vaccine over t years,
c: Infections averted over t years,
d: Courses of vaccine per infection averted (B / C)
}
\description{
format grid for heatmap plotting
}
