% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvw.R, R/onevax_xvwrh.R, R/parameters.R
\name{initial_params}
\alias{initial_params}
\alias{initial_params_xvw}
\alias{initial_params_xvwrh}
\title{Initial conditions for the model}
\usage{
initial_params_xvw(pars, coverage = 0)

initial_params_xvwrh(pars, coverage = 0, hes = 0)

initial_params(pars, n_vax = 1, coverage = 1)
}
\arguments{
\item{pars}{A parameter list containing `N0`, `q`, `prev_Asl` and `prev_Ash`
elements.}

\item{coverage}{a vector of length `n_vax` that sums to 1 denoting the
initial proportion in each vaccine stratum}

\item{hes}{proportion of population vaccine hesitant}

\item{n_vax}{an integer indicating the number of vaccine compartments}
}
\value{
A list of initial conditions

A list of initial conditions

A list of initial model states
}
\description{
Create initial conditions for the model

Create initial conditions for the model

Create initial conditions for the model
}
