% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_trial.R
\name{initial_params_trial}
\alias{initial_params_trial}
\title{Initial conditions for the model trial where the entire cohort is
in the high sexual activity group.}
\usage{
initial_params_trial(pars, n_vax = 1, p_v = 1)
}
\arguments{
\item{pars}{A parameter list containing `N0`, and `q` elements.}

\item{n_vax}{an integer indicating the number of vaccine compartments}

\item{p_v}{a vector of length `n_vax` that sums to 1 denoting the
initial proportion in each vaccine stratum}
}
\value{
A list of initial model states
}
\description{
Create initial conditions for the model trial
}
