% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xpvwrh.R
\name{initial_params_xpvwrh}
\alias{initial_params_xpvwrh}
\title{Initial conditions for the model}
\usage{
initial_params_xpvwrh(
  pars,
  coverage_p = 0,
  coverage_v = 0,
  hes = 0,
  t = FALSE,
  n_erlang = 1,
  n_diag_rec = 1
)
}
\arguments{
\item{pars}{A parameter list containing `N0`, `q`, `prev_Asl` and `prev_Ash`
elements.}

\item{coverage_p}{partial (one-dose) vaccine coverage of the population
already present (as a proportion)}

\item{coverage_v}{two-dose vaccine coverage of the population already
present (as a proportion)}

\item{hes}{proportion of population vaccine hesitant}

\item{t}{number of years, only use when using function outside of
run_onevax_xpvwrh() to generate initial conditions for tests}

\item{n_erlang}{integer giving the number of transitions that need to be
made through vaccine-protected strata until that protection has waned}

\item{n_diag_rec}{number of diagnosis history strata}
}
\value{
A list of initial conditions
}
\description{
Create initial conditions for the model
}
