% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvw_trial.R
\name{initial_params_xvw_trial}
\alias{initial_params_xvw_trial}
\title{Initial conditions for the model in a vaccine trial}
\usage{
initial_params_xvw_trial(pars, p_v = 0.5, n_erlang = 1)
}
\arguments{
\item{pars}{A parameter list containing `N0`, and `q` elements.}

\item{p_v}{scalar giving proportion of the trial cohort vaccinated}

\item{n_erlang}{integer giving the number of transitions that need to be
 made
through vaccine-protected strata until that protection has waned}
}
\value{
A list of initial conditions.
}
\description{
Create initial conditions for the model in a vaccine trial
}
