% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvw_trial.R
\name{initial_params_xvw_trial}
\alias{initial_params_xvw_trial}
\title{Initial conditions for the model in a vaccine trial}
\usage{
initial_params_xvw_trial(pars, p_v = 0.5)
}
\arguments{
\item{pars}{A parameter list containing `N0`, and `q` elements.}

\item{p_v}{scalar giving p_v of vaccination in trial cohort.}
}
\value{
A list of initial conditions.
}
\description{
Create initial conditions for the model in a vaccine trial
}
