% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvwrh.R
\name{initial_params_xvwrh}
\alias{initial_params_xvwrh}
\title{Initial conditions for the model}
\usage{
initial_params_xvwrh(pars, coverage = 0, hes = 0)
}
\arguments{
\item{pars}{A parameter list containing `N0`, `q`, `prev_Asl` and `prev_Ash`
elements.}

\item{coverage}{scalar giving initial coverage of vaccination in willing
population}

\item{hes}{proportion of population vaccine hesitant}
}
\value{
A list of initial conditions
}
\description{
Create initial conditions for the model
}
