% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_trial.R
\name{model_params_trial}
\alias{model_params_trial}
\title{Parameters for the vaccination trial model}
\usage{
model_params_trial(
  gono_params_trial = NULL,
  demographic_params_trial = NULL,
  initial_params_trial = NULL,
  vax_params = NULL,
  p_v = 0,
  n_erlang = 1,
  N = 6e+05,
  n_diag_rec = 1
)
}
\arguments{
\item{gono_params_trial}{A dataframe of natural history parameters}

\item{demographic_params_trial}{A dataframe of demographic parameters}

\item{initial_params_trial}{A list of starting conditions}

\item{vax_params}{A vector of vaccination params}

\item{p_v}{A scalar indicating the percentage of the trial cohort that
is vaccinated}

\item{n_erlang}{scalar giving the number of transitions that need to be made
through vaccine-protected strata until that protection has waned}

\item{N}{integer to assign the total number of individuals in the trial
(split equally across the two arms)}

\item{n_diag_rec}{integer giving the number of each X, V(erlang), and W stratum,
allowing tracking of diagnosis history. e.g for a n_diag_rec = 2 and erlang = 1,
there will be X.I, X.II, V1.I, V1.II, W.I, W.II strata. Where '.I'
corresponds to never-diagnosed individuals and '.II' is for individuals
diagnosed at least once.}
}
\value{
A list of inputs to the model many of which are fixed and
  represent data. These correspond largely to `user()` calls
  within the odin code, though some are also used in processing
  just before the model is run.
}
\description{
Parameters for the vaccination trial model
}
