% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvwrh.R
\name{restart_hes}
\alias{restart_hes}
\title{uses XVWRH model run in the absence of vaccination or hesitancy.
Saves down the number of individuals in each compartment, and moves
a given proportion (hes) of them from the X to the H strata to generate
new initial conditions in the presence of hesitancy.}
\usage{
restart_hes(y, hes = 0, branching = FALSE)
}
\arguments{
\item{y}{a transformed model run output}

\item{hes}{proportion of population vaccine hesitant}

\item{branching}{boolean to denote if xpvwrh branching model in use}
}
\value{
A list of initial conditions to restart a model with n_vax
vaccination levels, and a populated hestitant stratum in the given
proportion 'hes'
}
\description{
uses XVWRH model run in the absence of vaccination or hesitancy.
Saves down the number of individuals in each compartment, and moves
a given proportion (hes) of them from the X to the H strata to generate
new initial conditions in the presence of hesitancy.
}
