% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xpvwrh.R
\name{restart_hes}
\alias{restart_hes}
\title{uses XPVWRH model run in the absence of vaccination or hesitancy.
Saves down the number of individuals in each compartment, and moves
a given proportion (hes) of them from the X to the H strata to generate
new initial conditions in the presence of hesitancy.}
\usage{
restart_hes(
  y,
  n_vax = 6,
  hes = 0,
  n_erlang = 1,
  n_diag_rec = 1,
  branching = FALSE
)
}
\arguments{
\item{y}{a transformed model run output}

\item{hes}{proportion of population vaccine hesitant}

\item{n_erlang}{integer giving the number of transitions that need to be
made}

\item{n_diag_rec}{integer for the number of diagnosis history substrata}

\item{branching}{boolean to denote if xpvwrh branching model in use}
}
\value{
A list of initial conditions to restart a model with n_vax
vaccination levels, and a populated hestitant stratum in the given
proportion 'hes'
}
\description{
uses XPVWRH model run in the absence of vaccination or hesitancy.
Saves down the number of individuals in each compartment, and moves
a given proportion (hes) of them from the X to the H strata to generate
new initial conditions in the presence of hesitancy.
}
