% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novax.R
\name{run_novax_user}
\alias{run_novax_user}
\title{run model without vaccination for input parameter sets from
initialisation}
\usage{
run_novax_user(gono_params, tt)
}
\arguments{
\item{tt}{a numeric vector of times at which the model state is output}

\item{a}{set of gono_params}
}
\value{
A list of transformed model outputs
}
\description{
run model without vaccination for input parameter sets from
initialisation
}
