% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvwr.R
\name{run_onevax_xvwr}
\alias{run_onevax_xvwr}
\title{run model with single vaccine for input parameter sets, either from
initialisation or from equilibrium, those with waned vaccines are eligible
for revaccination (R), and return to the R stratum}
\usage{
run_onevax_xvwr(
  tt,
  gono_params,
  init_params = NULL,
  eff,
  dur,
  eff_revax = eff,
  dur_revax = dur,
  ve = 0,
  uptake = 0,
  strategy = "VbE",
  t_stop = 99
)
}
\arguments{
\item{tt}{a numeric vector of times at which the model state is output}

\item{gono_params}{list of gono params}

\item{init_params}{= NULL}

\item{eff}{scalar or numeric vector with same length as `gono_params` giving
efficacy of the vaccine (between 0-1)}

\item{dur}{scalar or numeric vector with same length as `gono_params`
giving duration of the vaccine (in years)}

\item{eff_revax}{scalar or numeric vector with same length as `gono_params`
giving efficacy of revaccination, default to same as primary}

\item{dur_revax}{scalar or numeric vector with same length as `gono_params`
giving duration of protection for revaccination, default to same as primary}

\item{ve}{scalar indicating % of population vaccinated before entry
(between 0-1)}

\item{uptake}{scalar or numeric vector with same length as `gono_params`
giving % of population vaccinated as part of strategy}

\item{strategy}{single character string in "VbE", "VoD(all)", "VoD(H)",
"VoA(all)", "VoA(H)", "VoD(L)+VoA(H)"}

\item{t_stop}{time at which vaccination should stop (years)}
}
\value{
A list of transformed model outputs
}
\description{
run model with single vaccine for input parameter sets, either from
initialisation or from equilibrium, those with waned vaccines are eligible
for revaccination (R), and return to the R stratum
}
