% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvwv.R
\name{run_onevax_xvwv}
\alias{run_onevax_xvwv}
\title{run model with single vaccine for input parameter sets, either from
initialisation or from equilibrium, those with waned vaccines are eligible
for revaccination, and return to the V compartment}
\usage{
run_onevax_xvwv(
  tt,
  gono_params,
  init_params = NULL,
  eff,
  dur,
  ve = 0,
  vd = 0,
  vs = 0,
  t_stop = 99
)
}
\arguments{
\item{tt}{a numeric vector of times at which the model state is output}

\item{gono_params}{a data frame of parameters}

\item{init_params}{= NULL}

\item{eff}{single numeric indicating efficacy of the vaccine (between 0-1)}

\item{dur}{single numeric indicating duration of the vaccine (in years)}

\item{ve}{single numeric indicating % of population vaccinated before entry
(between 0-1)}

\item{vd}{numeric indicating % of population vaccinated on diagnosis
(between 0-1) a scalar will apply to both activity groups, a vector of
length 2 will apply to each activity group separately}

\item{vs}{numeric indicating % of population vaccinated on screening
(between 0-1) a scalar will apply to both activity groups, a vector of
length 2 will apply to each activity group separately}

\item{t_stop}{time at which vaccination should stop (years)}
}
\value{
A list of transformed model outputs
}
\description{
run model with single vaccine for input parameter sets, either from
initialisation or from equilibrium, those with waned vaccines are eligible
for revaccination, and return to the V compartment
}
