% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvwv.R
\name{run_onevax_xvwv}
\alias{run_onevax_xvwv}
\title{run model with single vaccine for input parameter sets, either from
initialisation or from equilibrium, those with waned vaccines are eligible
for revaccination, and return to the V compartment}
\usage{
run_onevax_xvwv(
  tt,
  gono_params,
  init_params = NULL,
  eff,
  dur,
  ve = 0,
  uptake = 0,
  strategy = "VbE",
  t_stop = 99
)
}
\arguments{
\item{tt}{a numeric vector of times at which the model state is output}

\item{gono_params}{list of gono params}

\item{init_params}{= NULL}

\item{eff}{scalar or numeric vector with same length as `gono_params` giving
efficacy of the vaccine (between 0-1)}

\item{dur}{scalar or numeric vector with same length as `gono_params`
giving duration of the vaccine (in years)}

\item{ve}{scalar indicating pc of population vaccinated before entry
(between 0-1)}

\item{uptake}{scalar or numeric vector with same length as `gono_params`
giving pc of population vaccinated as part of strategy}

\item{strategy}{single character string in "VbE", "VoD", "VoD(H)",
"VoA", "VoA(H)", "VoD(L)+VoA(H)"}

\item{t_stop}{time at which vaccination should stop (years)}
}
\description{
run model with single vaccine for input parameter sets, either from
initialisation or from equilibrium, those with waned vaccines are eligible
for revaccination, and return to the V compartment
}
