% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trial.R
\name{run_trial}
\alias{run_trial}
\title{Run odin model of gonorrhoea vaccine trial with or without
vaccination}
\usage{
run_trial(
  tt,
  gono_params,
  init_params = NULL,
  vax_params = NULL,
  transform = TRUE,
  stochastic = FALSE
)
}
\arguments{
\item{tt}{a numeric vector of times at which the model state is output}

\item{gono_params}{a data frame of parameters}

\item{init_params}{= NULL}

\item{vax_params}{= NULL}

\item{transform}{= TRUE}

\item{stochastic}{logical indicating if the run should be made with the
default deterministic trial model in continuous time or stochastic trial
model in discrete time}
}
\description{
Run odin model of gonorrhoea vaccine trial with or without
vaccination
}
