% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_xpvwrh}
\alias{run_xpvwrh}
\title{Run odin model of gonorrhoea with or without vaccination}
\usage{
run_xpvwrh(
  tt,
  gono_params,
  demographic_params = NULL,
  init_params = NULL,
  vax_params = NULL,
  n_erlang = 1,
  n_diag_rec = 1,
  transform = TRUE
)
}
\arguments{
\item{tt}{a numeric vector of times at which the model state is output}

\item{gono_params}{A dataframe of natural history parameters}

\item{demographic_params}{A dataframe of demographic parameters}

\item{init_params}{A list of starting conditions}

\item{vax_params}{A vector of vaccination params}

\item{n_erlang}{integer giving the number of transitions that need to be
made through vaccine-protected strata until that protection has waned}

\item{n_diag_rec}{integer for the number of diagnosis history substrata}

\item{transform}{= TRUE}
}
\description{
Run odin model of gonorrhoea with or without vaccination
}
