% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{set_strategy}
\alias{set_strategy}
\title{Create mapping from vaccination strategy / uptake to model params}
\usage{
set_strategy(strategy, uptake)
}
\arguments{
\item{strategy}{one of "ve", "va", "vd" or "vt"}

\item{uptake}{single numeric between 0-1}
}
\value{
a list with elements vd and vs
}
\description{
Create mapping from vaccination strategy / uptake to model params
}
