% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xpvwrh.R
\name{stratum_index_xpvwrh}
\alias{stratum_index_xpvwrh}
\title{Generate the indices of all xpvwrh strata}
\usage{
stratum_index_xpvwrh(n_erlang)
}
\arguments{
\item{n_erlang}{integer giving the number of transitions that need to be
made through vaccine-protected strata until that protection has waned}
}
\value{
A list of strata with their indicies
}
\description{
Generate the indices of all xpvwrh strata
}
