% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratum_indexes.R
\name{stratum_index_xvwrh}
\alias{stratum_index_xvwrh}
\title{Generate the indices of all xvwrh trial strata}
\usage{
stratum_index_xvwrh(n_erlang = 1, n_diag_rec = 1, strategy = NULL)
}
\arguments{
\item{n_erlang}{integer giving the number of transitions that need to be
made through vaccine-protected strata until that protection has waned}

\item{n_diag_rec}{integer for the number of diagnosis history substrata}

\item{strategy}{string of vaccination strategy being considered}

}
\value{
A list of strata with their indicies
}
\description{
Generate the indices of all xvwrh trial strata
}
