% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novax.R
\name{vax_params0}
\alias{vax_params0}
\title{create vaccination parameters for use in novax model (null)}
\usage{
vax_params0()
}
\value{
A list parameters in the model input format
}
\description{
create vaccination parameters for use in novax model (null)
}
