% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_waning.R
\name{vax_params1w}
\alias{vax_params1w}
\title{currently waning is in stratum 5 (with strata 3:4 unused) for future
compatibility with dualvax, however may get performance boost by moving to
stratum 3}
\usage{
vax_params1w(dur = 1000, ve = 0, vs = 0, vd = 0, eff = 0)
}
\description{
currently waning is in stratum 5 (with strata 3:4 unused) for future
compatibility with dualvax, however may get performance boost by moving to
stratum 3
}
