% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_waning.R
\name{vax_params1w}
\alias{vax_params1w}
\title{create vaccination parameters for use in onevax_waning model}
\usage{
vax_params1w(eff = 0, dur = 1000, ve = 0, vs = 0, vd = 0)
}
\arguments{
\item{eff}{single numeric indicating efficacy of the vaccine (between 0-1)}

\item{dur}{single numeric indicating duration of the vaccine (in years)}

\item{ve}{single numeric indicating % of population vaccinated before entry
(between 0-1)}

\item{vs}{numeric indicating % of population vaccinated on screening
(between 0-1) a scalar will apply to both activity groups, a vector of
length 2 will apply to each activity group separately}

\item{vd}{numeric indicating % of population vaccinated on diagnosis
(between 0-1) a scalar will apply to both activity groups, a vector of
length 2 will apply to each activity group separately}
}
\value{
A list parameters in the model input format
currently waning is in stratum 5 (with strata 3:4 unused) for future
compatibility with dualvax, however may get performance boost by moving to
stratum 3
}
\description{
create vaccination parameters for use in onevax_waning model
}
