% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvw_trial.R
\name{vax_params_xvw_trial}
\alias{vax_params_xvw_trial}
\title{Create vaccination parameters for use in onevax_xvw_trial model,
assign who experiences vaccine effects, and how waning occurs.}
\usage{
vax_params_xvw_trial(
  vea = 0,
  vei = 0,
  ved = 0,
  ves = 0,
  dur = 1000,
  n_erlang = 1,
  stochastic = FALSE,
  n_diag_rec = 1,
  asymp_recorded = TRUE
)
}
\arguments{
\item{vea}{scalar indicating efficacy of the vaccine against acquisition
(between 0-1)}

\item{vei}{scalar indicating efficacy of the vaccine against infectiousness
(between 0-1)}

\item{ved}{scalar indicating efficacy of the vaccine against duration
(between 0-1)}

\item{ves}{scalar indicating efficacy of the vaccine against symptoms
(between 0-1)}

\item{dur}{scalar indicating duration of the vaccine (in years)}

\item{n_erlang}{integer giving the number of transitions that need to be
made through vaccine-protected strata until that protection has waned}

\item{stochastic}{logical indicating if the parameters are for the
default deterministic trial model in continuous time or stochastic trial
model in discrete time}

\item{n_diag_rec}{integer giving the number of each X, V(erlang), and W
stratum, allowing tracking of diagnosis history. e.g for a n_diag_rec = 2
and erlang = 1, there will be X.I, X.II, V1.I, V1.II, W.I, W.II strata.
Where '.I' corresponds to never-diagnosed individuals and '.II' is for
individuals diagnosed at least once.}

\item{asymp_recorded}{logical indicating if the trial screens for and
records asymptomatic diagnosis. If FALSE, asymptomatic infected individuals
undergoing treatment do not move diagnosis history stratum}
}
\value{
A list of parameters in the model input format
}
\description{
Create vaccination parameters for use in onevax_xvw_trial model,
assign who experiences vaccine effects, and how waning occurs.
}
