% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvw_trial.R
\name{vax_params_xvw_trial}
\alias{vax_params_xvw_trial}
\title{Create vaccination parameters for use in onevax_xvw_trial model,
assign who experiences vaccine effects, and how waning occurs.}
\usage{
vax_params_xvw_trial(
  vea = 0,
  vei = 0,
  ved = 0,
  ves = 0,
  dur = 1000,
  n_erlang = 1
)
}
\arguments{
\item{vea}{scalar indicating efficacy of the vaccine against acquisition
(between 0-1)}

\item{vei}{scalar indicating efficacy of the vaccine against infectiousness
(between 0-1)}

\item{ved}{scalar indicating efficacy of the vaccine against duration
(between 0-1)}

\item{ves}{scalar indicating efficacy of the vaccine against symptoms
(between 0-1)}

\item{dur}{scalar indicating duration of the vaccine (in years)}

\item{n_erlang}{scalar giving the number of transitions that need to be made
through vaccine-protected strata until that protection has waned}
}
\value{
A list of parameters in the model input format
}
\description{
Create vaccination parameters for use in onevax_xvw_trial model,
assign who experiences vaccine effects, and how waning occurs.
}
