% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvwr.R
\name{vax_params_xvwr}
\alias{vax_params_xvwr}
\title{create vaccination parameters for use in onevax_xvwr model}
\usage{
vax_params_xvwr(
  vea = 0,
  vei = 0,
  ved = 0,
  ves = 0,
  vea_revax = vea,
  vei_revax = vei,
  ved_revax = ved,
  ves_revax = ves,
  dur = 1000,
  dur_revax = dur,
  primary_uptake = 0,
  booster_uptake = primary_uptake,
  strategy = NULL,
  vbe = 0,
  t_stop = 99
)
}
\arguments{
\item{vea}{scalar indicating efficacy of the vaccine against acquisition
(between 0-1)}

\item{vei}{scalar indicating efficacy of the vaccine against infectiousness
(between 0-1)}

\item{ved}{scalar indicating efficacy of the vaccine against duration
(between 0-1)}

\item{ves}{scalar indicating efficacy of the vaccine against symptoms
(between 0-1)}

\item{vea_revax}{scalar indicating efficacy of revaccination against
acquisition (between 0-1)}

\item{vei_revax}{scalar indicating efficacy of revaccination against
infectiousness (between 0-1)}

\item{ved_revax}{scalar indicating efficacy of revaccination against
duration of infection (between 0-1)}

\item{ves_revax}{scalar indicating efficacy of revaccination against
symptoms (between 0-1)}

\item{dur}{scalar indicating duration of the vaccine (in years)}

\item{dur_revax}{duration of protection for revaccination,
default to same as primary}

\item{primary_uptake}{scalar or numeric vector with same length as
'gono_params' giving proportion of population undertaking primary
vaccination as part of strategy}

\item{booster_uptake}{scalar or numeric vector with same length as
'gono_params' giving proportion of population undertaking booster
vaccination after primary vaccination protection has waned}

\item{strategy}{single character string in "VoD", "VoD(H)",
"VoA", "VoA(H)", "VoD(L)+VoA(H)". Defaults to NULL i.e. no vaccination}

\item{vbe}{scalar indicating pc of population vaccinated before entry
(between 0-1)}

\item{t_stop}{time at which vaccination should stop (years)}
}
\value{
A list parameters in the model input format
}
\description{
create vaccination parameters for use in onevax_xvwr model
}
