% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevax_xvwv.R
\name{vax_params_xvwv}
\alias{vax_params_xvwv}
\title{create vaccination parameters for use in onevax_xvwv model}
\usage{
vax_params_xvwv(
  eff = 0,
  dur = 1000,
  uptake = 0,
  strategy = "VbE",
  ve = 0,
  t_stop = 99
)
}
\arguments{
\item{eff}{single numeric indicating efficacy of the vaccine (between 0-1)}

\item{dur}{single numeric indicating duration of the vaccine (in years)}

\item{uptake}{single numeric indicating % of population vaccinated as part
of strategy}

\item{strategy}{single character string in "VbE", "VoD(all)", "VoD(H)",
"VoA(all)", "VoA(H)", "VoD(L)+VoA(H)"}

\item{ve}{single numeric indicating % of population vaccinated before entry
(between 0-1)}

\item{t_stop}{time at which vaccination should stop (years)}
}
\value{
A list parameters in the model input format
}
\description{
create vaccination parameters for use in onevax_xvwv model
}
